+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to `iex-cloud-sdk` will be documented in this file
 
+## v0.1.0-beta.4 - 2019-10-02
+
+### Added
+
+##### Alternative Data
+
+- Crypto Price
+- Crypto Book
+- Crypto Symbols
+
+### Changed
+
+- Crypto Quote
+
 ## v0.1.0-beta.4 - 2019-10-01
 
 ### Added